feat: Add support for WindiCSS
Support WindiCSS color format in tailwind.config.js

- Moves the colors.js file into /colors/index.
- Moves the colorValues variable into its own file.
- Creates a new `colors/windi.js` file, which allows user to `require(‘daisyui/colors/windi’)`
