+++ colors/windi.js
@@ -1,0 +1,11 @@
+const colorValues = require('./color-values')
+
+let colorObject = {
+  "transparent": "transparent",
+}
+
+for (const [key, item] of Object.entries(colorValues)) {
+  colorObject[key] = `hsla(var(--${item.name}, ${item.value}) / var(--tw-bg-opacity, 1))`
+}
+
+module.exports = colorObject
